# Use concurrency to check mangled passwords