This is part of an attempt to improve crunchy's performance.

This PR focuses on improving the mangled password checks by using concurrency to speed up the search.
```
$ benchstat before after 
name                   old time/op  new time/op  delta
ValidatePassword-8      41.3s ± 1%   29.7s ± 1%  -28.12%  (p=0.000 n=8+8)
FoundInDictionaries-8   11.8s ± 1%    4.4s ± 1%  -62.24%  (p=0.000 n=8+8)
```


Hashing will be next